SEO_Automation_FullPack_ProflowLabs/seo_automation_code/requirements.txt
pandas
pyyaml
openpyxl
duckdb
requests
